python-requirements.txt
# Keep sorted

jsonschema==4.23.0
mypy==1.14.1
numpy==2.2.2
pylint==3.3.4
pyyaml==6.0.2
ruff==0.9.3
scikit-learn==1.6.1
scipy==1.15.1
sphinx==8.2.3
sphinxcontrib-spelling==8.0.1
sphinx-copybutton==0.5.2
sphinx-rtd-theme==3.0.2
tabulate==0.9.0
termcolor==2.5.0
